Split bundle() into smaller functions
