+++ src/wordpress-wasm/bundling/babel-plugin-react-fast-refresh.js
