+++ src/wordpress-wasm/bundling/babel-plugin-transpile-wordpress-jsx.ts
@@ -1,0 +1,9 @@
+import * as babel from '@babel/standalone';
+
+export default () => [
+	babel.availablePlugins['transform-react-jsx'],
+	{
+		pragma: 'window.wp.element.createElement',
+		pragmaFrag: 'Fragment',
+	},
+];
